lookup and remove from environment return object in pointer and thus calling functions now allocates pointers to pointers instead.
Before one had to allocate a new header_t on the heap to put the result from 'lookup/remove' in.
This (obviously) created a copy which could be used to destroy the 'object' and then the copy but the original still existed.
The solve is changed typedefinition of 'lookup/remove' so that they now takes a pointer to a pointer of an header_t as an argument to return the result in. This means that before calling 'lookup/remove' on has to allocate memory for a header_t* and after one is done with the header destroy the pointer pointer which refers to the header_t.

The tests have been changed accordingly to the new specification and they executes to perfection without any memory-issues.